move client data frame indexer and builder into own files
